Change orderContext to allow items to be added to cart without user being logged in
