+++ src/components/header/index.js
@@ -1,14 +1,12 @@
 import { Link } from "gatsby"
 import PropTypes from "prop-types"
-import React, { useContext } from "react"
+import React from "react"
 import DrawerMenu from "../drawerMenu"
 import LocationBanner from "./locationBanner"
 import Cart from "../cart"
-import { AuthContext } from "../../context/authContext"
 import { HeaderWrapper, HeaderContent, HeaderLogo } from "./styledComponents"
 
 const Header = () => {
-  const { loggedIn } = useContext(AuthContext)
   return (
     <HeaderWrapper>
       <HeaderContent>
@@ -18,7 +16,7 @@ const Header = () => {
             <b>SCD</b>
           </HeaderLogo>
         </Link>
-        {loggedIn ? <Cart /> : <Link to={"/login"}>Login</Link>}
+        <Cart />
       </HeaderContent>
       <LocationBanner />
     </HeaderWrapper>
