+++ src/components/productCard/addToCartButton/index.js
@@ -1,10 +1,11 @@
 import React, { useContext } from "react"
-import { OrderContext } from "../../../context/orderContext"
+import OrderContext from "../../../context/orderContext"
 import { StyledAddToCartButton } from "./styledComponents"
 import addToCartIcon from "../../../images/add-to-cart.svg"
 
-const AddToCartButton = ({ price, productId }) => {
+export default ({ price, productId }) => {
   const { addToOrder } = useContext(OrderContext)
+
   return (
     <StyledAddToCartButton onClick={() => addToOrder(productId, 1)}>
       <img src={addToCartIcon} alt="Add to cart" />
@@ -12,5 +13,3 @@ const AddToCartButton = ({ price, productId }) => {
     </StyledAddToCartButton>
   )
 }
-
-export default AddToCartButton
