+++ src/hooks/useConcentrates.js
@@ -13,27 +13,32 @@ export default () => {
         ) {
           edges {
             node {
-              type
-              category {
-                name
-              }
-              subCategory {
-                name
-              }
-              description
-              featured
               id
+              productId
+              name
+              brand {
+                name
+              }
               image {
                 url
               }
-              name
+              strain {
+                indicaSativaRatio
+                name
+              }
               variants {
-                price
-                productId
-                unitWeight
-                unit
+                id
                 flavor
-                description
+                price
+              }
+              type
+              category {
+                id
+                name
+              }
+              subCategory {
+                id
+                name
               }
             }
           }
